requirements.txt
Django==1.5.2
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
djangorestframework==2.3.7
gunicorn==17.5
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
